backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "eleanor"
version = "1.0.0-alpha"
description = "Open-source DFIR platform with Sentinel-style UI and built-in case management"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.12"
authors = [
    { name = "Eleanor Contributors" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: FastAPI",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
]

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.1",
    "pydantic[email]>=2.6.0",
    "pydantic-settings>=2.1.0",
    "email-validator>=2.1.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",
    "elasticsearch[async]>=8.12.0",
    "aiohttp>=3.9.0",
    "redis>=5.0.1",
    "httpx>=0.26.0",
    "python-ldap>=3.4.4",
    "authlib>=1.3.0",
    "aiofiles>=23.2.1",
    "python-magic>=0.4.27",
    "grpcio>=1.60.0",
    "pyvelociraptor>=0.1.9",
    # Celery task queue
    "celery>=5.3.0",
    # Dissect forensic parsers
    "dissect.target>=3.0",
    "dissect.ntfs>=3.0",
    "dissect.regf>=3.0",
    "dissect.esedb>=3.0",
    "dissect.util>=3.0",
    # Network capture parsing
    "scapy>=2.5.0",
    # Windows Event Log parsing
    "python-evtx>=0.7.4",
    # Memory forensics
    "volatility3>=2.5.0",
    # Cloud Storage SDKs
    "boto3>=1.34.0",  # AWS S3
    "azure-storage-blob>=12.19.0",  # Azure Blob Storage
    "azure-identity>=1.15.0",  # Azure authentication
    "google-cloud-storage>=2.14.0",  # Google Cloud Storage
    # Microsoft Integrations
    "msal>=1.26.0",  # Microsoft Authentication Library
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
    "black>=24.1.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

[project.urls]
Homepage = "https://github.com/eleanor-dfir/eleanor"
Documentation = "https://github.com/eleanor-dfir/eleanor/docs"
Repository = "https://github.com/eleanor-dfir/eleanor"
Issues = "https://github.com/eleanor-dfir/eleanor/issues"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
# Parser registry imports modules for side effects (registration)
"app/parsers/registry.py" = ["F401"]
# Parser __init__ re-exports parsers
"app/parsers/formats/__init__.py" = ["F401"]
# Parser format files import for availability checks and may have camelcase library imports
"app/parsers/formats/*.py" = ["F401", "F821", "N813"]
# Connectors may have unused variables during development
"app/connectors/*.py" = ["F841"]
# Base exception classes don't need Error suffix
"app/exceptions.py" = ["N818"]
# Rate limit exception is intentionally named
"app/middleware/rate_limit.py" = ["N818"]
# RBAC model has forward references
"app/models/rbac.py" = ["F821"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
